update loadable-component example
