+++ examples/with-loadable-component/.babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["razzle/babel"],
+  "plugins": ["@loadable/babel-plugin"]
+}
