+++ examples/with-loadable-component/public/favicon.ico
