+++ examples/with-loadable-component/public/robots.txt
