+++ examples/with-loadable-component/src/BodyPart.js
