+++ examples/with-loadable-component/src/Footer.js
