+++ examples/with-loadable-component/src/index.js
