--- examples/with-loadable-components/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["razzle/babel"],
-  "plugins": ["loadable-components/babel"]
-}
