--- examples/with-loadable-components/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "razzle-examples-with-loadable-components",
-  "version": "3.0.0-alpha.2",
-  "license": "MIT",
-  "scripts": {
-    "start": "razzle start",
-    "build": "razzle build",
-    "test": "razzle test --env=jsdom",
-    "start:prod": "NODE_ENV=production node build/server.js"
-  },
-  "dependencies": {
-    "express": "^4.16.3",
-    "loadable-components": "^1.4.0",
-    "razzle": "^3.0.0-alpha.2",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0"
-  }
-}
