--- examples/with-react-loadable/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["razzle/babel"],
-  "plugins": ["react-loadable/babel"]
-}
