--- examples/with-react-loadable/.gitignore
@@ -1,12 +1,0 @@
-logs
-*.log
-npm-debug.log*
-.DS_Store
-
-coverage
-node_modules
-build
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
