--- examples/with-react-loadable/README.md
@@ -1,20 +1,0 @@
-# Razzle React Loadable Example
-
-## How to use
-Download the example [or clone the whole project](https://github.com/jaredpalmer/razzle.git):
-
-```bash
-curl https://codeload.github.com/jaredpalmer/razzle/tar.gz/master | tar -xz --strip=2 razzle-master/examples/with-react-loadable
-cd with-react-loadable
-```
-
-Install it and run:
-
-```bash
-yarn install
-yarn start
-```
-
-## Idea behind the example
-This sample has the full SSR setup as described in the [react-loadable docs](https://github.com/thejameskyle/react-loadable#--------------server-side-rendering).
-It demonstrates simple dynamic imports using [react-loadable](https://github.com/thejameskyle/react-loadable) on top of SSR with razzle.
