--- examples/with-react-loadable/public/robots.txt
@@ -1,2 +1,0 @@
-User-agent: *
-
