--- examples/with-react-loadable/razzle.config.js
@@ -1,19 +1,0 @@
-const { ReactLoadablePlugin } = require('react-loadable/webpack');
-
-module.exports = {
-  modify: (config, { target }) => {
-    if (target === 'web') {
-      return {
-        ...config,
-        plugins: [
-          ...config.plugins,
-          new ReactLoadablePlugin({
-            filename: './build/react-loadable.json',
-          }),
-        ],
-      };
-    }
-
-    return config;
-  },
-};
