--- examples/with-react-loadable/src/App.test.js
@@ -1,16 +1,0 @@
-import App from './App';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import MemoryRouter from 'react-router-dom/MemoryRouter';
-
-describe('<App />', () => {
-  test('renders without exploding', () => {
-    const div = document.createElement('div');
-    ReactDOM.render(
-      <MemoryRouter>
-        <App />
-      </MemoryRouter>,
-      div
-    );
-  });
-});
