--- examples/with-react-loadable/src/Intro.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const Intro = () => (
-  <p className="Home-intro">
-    To get started, edit <code>src/App.js</code> or <code>src/Home.js</code> and
-    save to reload.
-  </p>
-);
-
-export default Intro;
