--- examples/with-react-loadable/src/Logo.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import logo from './react.svg';
-
-const Logo = () => <img src={logo} className="Home-logo" alt="logo" />;
-
-export default Logo;
