--- examples/with-react-loadable/src/Welcome.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Welcome = () => <h2>Welcome to Razzle</h2>;
-
-export default Welcome;
